// src/utils/styles/constants.d.ts
export declare const symbols: {
    STEP_ACTIVE: string;
    STEP_CANCEL: string;
    STEP_ERROR: string;
    STEP_SUBMIT: string;
    BAR_START: string;
    BAR: string;
    BAR_END: string;
    RADIO_ACTIVE: string;
    RADIO_INACTIVE: string;
    CHECKBOX_ACTIVE: string;
    CHECKBOX_SELECTED: string;
    CHECKBOX_INACTIVE: string;
    PASSWORD_MASK: string;
    BAR_H: string;
    CORNER_TOP_RIGHT: string;
    CONNECT_LEFT: string;
    CORNER_BOTTOM_RIGHT: string;
    INFO: string;
    SUCCESS: string;
    WARN: string;
    ERROR: string;
};
//# sourceMappingURL=constants.d.ts.map